+++ emav/probe_w.3.c
@@ -323,7 +323,7 @@ z_surge(float *zl, int na)
 
     z_chr(zc);
     z0 = zc[0];
-    R = (unsigned long)((double *) calloc(na * 2, sizeof(double)));
+    R = (unsigned long)((double *) calloc((size_t)(na * 2), sizeof(double)));
     c_ones(R, na);
     fd_window(R, na);
     w0 = mean_real(R, na);
@@ -367,14 +367,14 @@ thev_init()
 {
     int m;
 
-    qq1 = (float *) calloc(MAX_NF * 2 * 2, sizeof(float));
+    qq1 = (float *) calloc((size_t)(MAX_NF * 2 * 2), sizeof(float));
     if (qq1 == NULL) {
         decide(0, 1, "Can't allocate thev memory (qq1)");
         return (0);
     }
 
     os = MAX_NF * 2 * 3;	//allow for ps, zs, px
-    pzs = (unsigned long)((float *) calloc(os * 2, sizeof(float)));
+    pzs = (unsigned long)((float *) calloc((size_t)(os * 2), sizeof(float)));
     if (pzs == NULL) {
         decide(0, 1, "Can't allocate thev memory (pzs)");
 	free(qq1);
@@ -383,7 +383,7 @@ thev_init()
 
     m = (nthc < 6) ? 6 : nthc;	//allow for pl, zl, pf, z0, pr, cd
     ol = MAX_NF * 2 * m;;
-    pzl = (unsigned long)((float *) calloc(ol * 2, sizeof(float)));
+    pzl = (unsigned long)((float *) calloc((size_t)(ol * 2), sizeof(float)));
     if (pzl == NULL) {
         decide(0, 1, "Can't allocate thev memory (pzl)");
 	free(qq1);
@@ -1226,7 +1226,7 @@ cavity_length(float *pc, int nf, double vs)
     int i, ii, ir, i1, i2, m, n;
 
     n = nf * 2 + 2;
-    p = (unsigned long)((float *) calloc(n, sizeof(float)));
+    p = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
     for (i = 0; i < nf; i++) {
 	ir = 2 * i;
 	ii = 2 * i + 1;
@@ -1749,9 +1749,9 @@ thev_adjust(char *cfn, char *sfn, SAV_FFT *pft, int sil)
 	    free_null(iadj);
 	    free_null(refl);
 	    n = (nthf < pft[j].npts) ? nthf : pft[j].npts;
-	    padj = (unsigned long)((float *) calloc(n, sizeof(float)));
-	    iadj = (unsigned long)((float *) calloc(n, sizeof(float)));
-	    refl = (unsigned long)((float *) calloc(n, sizeof(float)));
+	    padj = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
+	    iadj = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
+	    refl = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
 	    for (i = 0; i < n; i++) {
 		ir = 2 * i;
 		ii = 2 * i + 1;
